requirements.txt
Django==1.11.3
django-cors-headers==2.1.0
djangorestframework==3.6.3
psycopg2==2.7.1
pytz==2017.2
